agentBackend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.11
fastapi==0.115.6
fastavro==1.10.0
filelock==3.17.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
packaging==24.2
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
